requirements.txt
Flask==3.0.2
opencv-python
numpy
scikit-image
scipy
imagehash
Pillow